Review nits
