+++ crates/optimism/txpool/src/supervisor/mod.rs
@@ -9,3 +9,5 @@ mod errors;
 pub use errors::{InteropTxValidatorError, InvalidInboxEntry};
 mod message;
 pub use message::ExecutingDescriptor;
+mod validator;
+pub use validator::is_valid_cross_tx;
